191112 | User-Article & User-Comment Relation

+++ 05_Django/05_django_form/articles/views.py
@@ -7,7 +7,7 @@ from IPython import embed
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
 import hashlib
- 
+  
 # Create your views here.
 def index(request):
   #embed()
@@ -39,14 +39,9 @@ def create(request):
     # 인스턴스에 데이터를 채워서, 유효성 검증을 진행한다
     form = ArticleForm(request.POST)
     if form.is_valid():
-      # form이 dict형태로 바뀌어져서 들어옴
-      # cleaned_data를 통해 디셔너리 안 데이터를 검증한다
-      # 1. Form Class
-      # title = form.cleaned_data.get('title')
-      # content = form.cleaned_data.get('content')
-      # article = Article.objects.create(title=title, content=content)
-      # 2. Model Form Class
-      article = form.save()
+      article = form.save(commit=False)
+      article.user = request.user
+      article.save()
     return redirect('articles:detail', article.pk)
   else:
     form = ArticleForm()
@@ -77,60 +72,74 @@ def detail(request, article_pk):
 def delete(request, article_pk):
   if request.user.is_authenticated:
     article = get_object_or_404(Article, pk=article_pk)
+    if request.user == article.user:
     # if request.method == 'POST':
     #   article.delete()
     #   return redirect('articles:index')
     # else:
     #   return redirect('articles:detail', article.pk)
-    article.delete()
+      article.delete()
+    else:
+      return redirect('articles:detail', article.pk)
   return redirect('articles:index')
 
 @login_required
 def update(request, article_pk):
   article= get_object_or_404(Article, pk=article_pk)
-  if request.method == 'POST':
-    form = ArticleForm(request.POST, instance=article)
-    if form.is_valid():
-      # article.title = form.cleaned_data('title')
-      # article.content = form.cleaned_data('content')
-      article = form.save()
-      return redirect('articles:detail', article.pk)
+  if request.user == article.user:
+    if request.method == 'POST':
+      form = ArticleForm(request.POST, instance=article)
+      if form.is_valid():
+        # article.title = form.cleaned_data('title')
+        # article.content = form.cleaned_data('content')
+        article = form.save()
+        return redirect('articles:detail', article.pk)
+    else:
+      # form = ArticleForm(initial={
+      #   'title':article.title,
+      #   'content':article.content
+      # })
+      form = ArticleForm(instance=article)
+      # 2가지 form 형식
+      # 1. GET -> 초기값을 폼에 넣어서 사용자에게 던져줌
+      # 2. POST -> is_valid가 False가 리턴됐을 때, 오류 메시지 포함해서 사용자에게 던져줌
   else:
-    # form = ArticleForm(initial={
-    #   'title':article.title,
-    #   'content':article.content
-    # })
-    form = ArticleForm(instance=article)
-    # 2가지 form 형식
-    # 1. GET -> 초기값을 폼에 넣어서 사용자에게 던져줌
-    # 2. POST -> is_valid가 False가 리턴됐을 때, 오류 메시지 포함해서 사용자에게 던져줌
+    return redirect('articles:index')
 
-    context = {
-      'form':form, 
-      'article':article 
-    }
-    #return render(request, 'articles/create.html', context)
-    return render(request, 'articles/form.html', context)
+  context = {
+    'form':form, 
+    'article':article 
+  }
+  #return render(request, 'articles/create.html', context)
+  return render(request, 'articles/form.html', context)
 
 @require_POST
 def comments_create(request, article_pk):
   if request.user.is_authenticated:
-    article = get_object_or_404(Article, pk=article_pk)
+    #article = get_object_or_404(Article, pk=article_pk)
     comment_form = CommentForm(request.POST)
     if comment_form.is_valid():
       # save() 메서드 -> 선택 인자 : (기본값) commit=True
       # comment의 article정보를 넣어줘야 하는데 정보가 없다면 일단 막아줘야한다
       # commit=False 하면 DB에 바로 저장되는 것을 막아준다. 객체 하나만 만들어진 상태
       comment = comment_form.save(commit=False)
-      comment.article = article
+      comment.user = request.user
+      # instance를 바로 넣어주기
+      #comment.article = article
+      # django가 만든 db형식에 맞추어 넣어주기
+      comment.article_id = article_pk
       comment.save()
-  return redirect('articles:detail', article.pk)
+  #return redirect('articles:detail', article.pk)
+  return redirect('articles:detail', article_pk)
 
 
 
 @require_POST
 def comments_delete(request, article_pk, comment_pk):
+  # 1. 로그인 여부 확인
   if request.user.is_authenticated:
     comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
+    # 2. 로그인한 사용자와 댓글 작성자가 같은 경우
+    if request.user == comment.user:
+      comment.delete()
   return redirect('articles:detail', article_pk)
